+++ tests/unit/test-class-wc-payment-gateway-wcpay.php
@@ -21,8 +21,8 @@ use WCPay\Duplicate_Payment_Prevention_Service;
 use WCPay\Exceptions\Amount_Too_Small_Exception;
 use WCPay\Exceptions\API_Exception;
 use WCPay\Fraud_Prevention\Fraud_Prevention_Service;
-use WCPay\Internal\Payment\Routing\Factor;
-use WCPay\Internal\Payment\Routing\Router;
+use WCPay\Internal\Payment\Factor;
+use WCPay\Internal\Payment\Router;
 use WCPay\Internal\Service\PaymentProcessingService;
 use WCPay\Payment_Information;
 use WCPay\WooPay\WooPay_Utilities;
